+++ contracts/interfaces/IBondPosition.sol
@@ -30,12 +30,12 @@ interface IBondPosition is IERC721Metadata, IERC721Permit, IBondPositionEvents {
     function mint(
         address asset,
         uint256 amount,
-        uint256 lockEpochs,
+        uint16 lockEpochs,
         address recipient,
         bytes calldata data
     ) external returns (uint256);
 
-    function adjustPosition(uint256 tokenId, int256 amount, int256 lockEpochs, bytes calldata data) external;
+    function adjustPosition(uint256 tokenId, int256 amount, int16 lockEpochs, bytes calldata data) external;
 
     function burnExpiredPosition(uint256 tokenId) external;
